Ass3_Career_Advisory/requirements.txt
durable-rules==2.0.28
